Create ImportSampleFileUseCase
