+++ src/domain/entities/metadata/DataSet.ts
@@ -1,11 +1,13 @@
 export interface DataSet {
     id: string;
     name: string;
-    dataElements: {
-        id: string;
-        name: string;
-        code: string;
-        categoryCombo: { id: string };
-    }[];
+    dataElements: DataElement[];
     categoryCombo: string;
+}
+
+export interface DataElement {
+    id: string;
+    name: string;
+    code: string;
+    categoryCombo: { id: string };
 }
